# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GenZVerse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface User {
  id: number;
  email: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  occupation?: string | null;
  interests: string[];
  goals: string[];
  /** @nullable */
  bio?: string | null;
  onboardingCompleted: boolean;
  lifeScore: number;
  xp: number;
  level: number;
  streak: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface SignupInput {
  email: string;
  password: string;
  fullName: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserUpdate {
  username?: string;
  fullName?: string;
  avatarUrl?: string;
  age?: number;
  occupation?: string;
  bio?: string;
  interests?: string[];
  goals?: string[];
}

export interface OnboardingInput {
  username: string;
  age?: number;
  occupation?: string;
  interests: string[];
  goals: string[];
}

export interface DashboardStats {
  lifeScore: number;
  lifeScoreTrend: number;
  productivityScore: number;
  productivityTrend: number;
  socialScore: number;
  socialTrend: number;
  learningScore: number;
  learningTrend: number;
  financeScore: number;
  financeTrend: number;
  styleScore: number;
  styleTrend: number;
  streak: number;
  xp: number;
  level: number;
  tasksCompleted: number;
  squadsJoined: number;
  challengesWon: number;
}

export interface Squad {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  memberCount: number;
  onlineCount: number;
  xp: number;
  category: string;
  /** @nullable */
  avatarUrl?: string | null;
  isFeatured: boolean;
  createdAt: string;
}

export interface SquadInput {
  name: string;
  description?: string;
  category: string;
}

export interface Community {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  memberCount: number;
  category: string;
  /** @nullable */
  avatarUrl?: string | null;
  isFeatured: boolean;
  tags: string[];
  createdAt: string;
}

export interface Challenge {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  xpReward: number;
  category: string;
  difficulty: string;
  participantCount: number;
  daysLeft: number;
  isActive: boolean;
  createdAt: string;
}

export interface Activity {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  xpEarned?: number | null;
  createdAt: string;
}