# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GenZVerse API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: dashboard
    description: Dashboard data
  - name: squads
    description: Social squads
  - name: communities
    description: Communities
  - name: challenges
    description: Challenges
  - name: activities
    description: Activity feed
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me:
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete onboarding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard life scores and stats
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Get recent activity feed for dashboard
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

  /squads:
    get:
      operationId: getSquads
      tags: [squads]
      summary: List squads
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Squad"
    post:
      operationId: createSquad
      tags: [squads]
      summary: Create a squad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SquadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Squad"

  /squads/featured:
    get:
      operationId: getFeaturedSquads
      tags: [squads]
      summary: Get featured squads
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Squad"

  /squads/{id}:
    get:
      operationId: getSquad
      tags: [squads]
      summary: Get a squad by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Squad"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /communities:
    get:
      operationId: getCommunities
      tags: [communities]
      summary: List communities
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Community"

  /communities/featured:
    get:
      operationId: getFeaturedCommunities
      tags: [communities]
      summary: Get featured communities
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Community"

  /challenges:
    get:
      operationId: getChallenges
      tags: [challenges]
      summary: List challenges
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"

  /activities:
    get:
      operationId: getActivities
      tags: [activities]
      summary: Get activity feed
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        occupation:
          type: ["string", "null"]
        interests:
          type: array
          items:
            type: string
        goals:
          type: array
          items:
            type: string
        bio:
          type: ["string", "null"]
        onboardingCompleted:
          type: boolean
        lifeScore:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - email
        - interests
        - goals
        - onboardingCompleted
        - lifeScore
        - xp
        - level
        - streak
        - createdAt

    SignupInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
      required:
        - email
        - password
        - fullName

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    UserUpdate:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: string
        age:
          type: integer
        occupation:
          type: string
        bio:
          type: string
        interests:
          type: array
          items:
            type: string
        goals:
          type: array
          items:
            type: string

    OnboardingInput:
      type: object
      properties:
        username:
          type: string
        age:
          type: integer
        occupation:
          type: string
        interests:
          type: array
          items:
            type: string
        goals:
          type: array
          items:
            type: string
      required:
        - username
        - interests
        - goals

    DashboardStats:
      type: object
      properties:
        lifeScore:
          type: integer
        lifeScoreTrend:
          type: integer
        productivityScore:
          type: integer
        productivityTrend:
          type: integer
        socialScore:
          type: integer
        socialTrend:
          type: integer
        learningScore:
          type: integer
        learningTrend:
          type: integer
        financeScore:
          type: integer
        financeTrend:
          type: integer
        styleScore:
          type: integer
        styleTrend:
          type: integer
        streak:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        tasksCompleted:
          type: integer
        squadsJoined:
          type: integer
        challengesWon:
          type: integer
      required:
        - lifeScore
        - lifeScoreTrend
        - productivityScore
        - productivityTrend
        - socialScore
        - socialTrend
        - learningScore
        - learningTrend
        - financeScore
        - financeTrend
        - styleScore
        - styleTrend
        - streak
        - xp
        - level
        - tasksCompleted
        - squadsJoined
        - challengesWon

    Squad:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        memberCount:
          type: integer
        onlineCount:
          type: integer
        xp:
          type: integer
        category:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - memberCount
        - onlineCount
        - xp
        - category
        - isFeatured
        - createdAt

    SquadInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
      required:
        - name
        - category

    Community:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        memberCount:
          type: integer
        category:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - memberCount
        - category
        - isFeatured
        - tags
        - createdAt

    Challenge:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        xpReward:
          type: integer
        category:
          type: string
        difficulty:
          type: string
        participantCount:
          type: integer
        daysLeft:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - xpReward
        - category
        - difficulty
        - participantCount
        - daysLeft
        - isActive
        - createdAt

    Activity:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        xpEarned:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - type
        - description
        - createdAt