# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GenZVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: number;
  type: string;
  description: string;
  /** @nullable */
  xpEarned?: number | null;
  createdAt: string;
}