# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GenZVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Challenge {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  xpReward: number;
  category: string;
  difficulty: string;
  participantCount: number;
  daysLeft: number;
  isActive: boolean;
  createdAt: string;
}