# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GenZVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  lifeScore: number;
  lifeScoreTrend: number;
  productivityScore: number;
  productivityTrend: number;
  socialScore: number;
  socialTrend: number;
  learningScore: number;
  learningTrend: number;
  financeScore: number;
  financeTrend: number;
  styleScore: number;
  styleTrend: number;
  streak: number;
  xp: number;
  level: number;
  tasksCompleted: number;
  squadsJoined: number;
  challengesWon: number;
}