# lib/api-zod/src/generated/types/onboardingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GenZVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingInput {
  username: string;
  age?: number;
  occupation?: string;
  interests: string[];
  goals: string[];
}