# lib/api-zod/src/generated/types/squad.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GenZVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Squad {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  memberCount: number;
  onlineCount: number;
  xp: number;
  category: string;
  /** @nullable */
  avatarUrl?: string | null;
  isFeatured: boolean;
  createdAt: string;
}