# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GenZVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  occupation?: string | null;
  interests: string[];
  goals: string[];
  /** @nullable */
  bio?: string | null;
  onboardingCompleted: boolean;
  lifeScore: number;
  xp: number;
  level: number;
  streak: number;
  createdAt: string;
}