# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GenZVerse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  username?: string;
  fullName?: string;
  avatarUrl?: string;
  age?: number;
  occupation?: string;
  bio?: string;
  interests?: string[];
  goals?: string[];
}